+++ R/R6Class-BipartiteSBM_sampler.R
@@ -12,15 +12,16 @@ BipartiteSBM_sampler <-
       #' @param nbNodes number of nodes in the network
       #' @param blockProp parameters for block proportions (vector of list of vectors)
       #' @param connectParam list of parameters for connectivity with a matrix of means 'mean' and an optional scalar for the variance 'var'. The dimensions of mu must match \code{blockProp} lengths
+      #' @param dimLabels optional labels of each dimension (in row, in column)
       #' @param covarParam optional vector of covariates effect
       #' @param covarList optional list of covariates data
-      initialize = function(model, nbNodes, blockProp, connectParam, covarParam=numeric(0), covarList=list()) {
+      initialize = function(model, nbNodes, blockProp, connectParam, dimLabels=list(row="rowLabel", col="colLabel"), covarParam=numeric(0), covarList=list()) {
         ## SANITY CHECKS
         stopifnot(length(blockProp) ==  2,
                   length(blockProp[[1]]) ==  nrow(connectParam$mean),
                   length(blockProp[[2]]) ==  ncol(connectParam$mean))
         names(blockProp) <- c("row", "col")
-        super$initialize(model, nbNodes, blockProp, connectParam, covarParam, covarList)
+        super$initialize(model, nbNodes, blockProp, dimLabels, connectParam, covarParam, covarList)
         self$rIncidence()
       },
       #' @description a method to generate a vector of block indicators
@@ -52,6 +53,8 @@ BipartiteSBM_sampler <-
       nbBlocks    = function(value) {sapply(private$pi, length)},
       #' @field nbDyads number of dyads (potential edges in the network)
       nbDyads     = function(value) {private$dim[1] * private$dim[2]},
+      #' @field nbConnectParam number of parameter used for the connectivity
+      nbConnectParam = function(value) {self$nbBlocks^2},
       #' @field memberships list of size 2: vector of memberships in row, in column.
       memberships = function(value) {lapply(private$Z, as_clustering)},
       #' @field indMemberships list of 2 matrix for clustering memberships