+++ NEWS.md
@@ -1,3 +1,9 @@
+# sbm 0.2.2 - minor release
+
+* added fields dimLabels and covarArray for class SBM
+* added fields nbConnectParam, penalty and entropy  for simple and bipartite SBM (sampler and fit)
+* changing roundProduct function to work on list of covariates
+
 # sbm 0.2.1 - minor release
 
 * various bug fixes, especially for directed network in Simple SBM, covariates for Bipartite LBM